收藏商品、发布模块完成

+++ app/src/main/java/com/triplebro/domineer/graduationdesignproject/activities/SubmitActivity.java
@@ -84,14 +84,14 @@ public class SubmitActivity extends Activity implements View.OnClickListener, On
                 submitInfo.setNickname(userInfo.getString("nickname", ""));
                 submitInfo.setUser_head(userInfo.getString("userHead", ""));
                 submitInfo.setSubmit_content(et_submit_content.getText().toString().trim());
-                long submit_id = submitManager.UploadSubmitInfo(submitInfo);
+                int submit_id = submitManager.UploadSubmitInfo(submitInfo);
                 List<String> data = submitAdapter.getData();
                 if (data.size() > 0 && data.size() < 9) {
                     data.remove(data.size()-1);
                 }
                 for (String s :data) {
                     SubmitImageInfo submitImageInfo = new SubmitImageInfo();
-                    submitImageInfo.setSubmit_id((int) submit_id);
+                    submitImageInfo.setSubmit_id(submit_id);
                     submitImageInfo.setSubmit_image(s);
                     submitManager.UploadSubmitImageInfo(submitImageInfo);
                 }
